save/load tcdb

+++ include/article.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <fstream>
 #include <string>
 #include <vector>
 
@@ -6,14 +7,14 @@ class Article {
 protected:
   std::string name;
   std::string manufacturer;
-  int stock;
-  int diameter;
+  int64_t stock;
+  int64_t diameter;
   float price;
   char type;
 
 public:
   Article();
-  Article(std::string, std::string, int, int, float, char);
+  Article(std::string, std::string, int64_t, int64_t, float, char);
   virtual ~Article();
 
   std::string getName(void);
@@ -22,11 +23,11 @@ public:
   std::string getManufacturer(void);
   void setManufacturer(std::string);
 
-  int getStock(void);
-  void setStock(int);
+  int64_t getStock(void);
+  void setStock(int64_t);
 
-  int getDiameter(void);
-  void setDiameter(int);
+  int64_t getDiameter(void);
+  void setDiameter(int64_t);
 
   float getPrice(void);
   void setPrice(float);
