+++ include/company.hpp
@@ -6,18 +6,18 @@
 class Company : public Customer {
 private:
   std::string vat;
-  int volumeDiscount;
+  int64_t volumeDiscount;
 
 public:
   Company();
-  Company(std::string, std::string, char, std::string, int);
+  Company(std::string, std::string, char, std::string, int64_t);
   ~Company();
 
   std::string getVAT(void);
   void setVAT(std::string);
 
-  int getVolumeDiscount(void);
-  void setVolumeDiscount(int);
+  int64_t getVolumeDiscount(void);
+  void setVolumeDiscount(int64_t);
 
   void show(void);
 };
